# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Nitzutz Spark AI Mentoring API
  description: |
    # 🚀 Nitzutz Spark API
    
    Enterprise-grade AI-powered personal learning and mentoring platform.
    
    ## 🔑 Authentication
    Currently in single-user mode. Multi-user JWT authentication coming soon.
    
    ## 📊 Rate Limits
    - **General API**: 1000 requests per 15 minutes
    - **AI Endpoints**: 30 requests per 15 minutes
    - **Auth Endpoints**: 10 requests per 15 minutes
    
    ## 🆘 Support
    - **Email**: support@nitzutz-spark.com
    - **Docs**: https://docs.nitzutz-spark.com
    - **GitHub**: https://github.com/nitzutz-spark
    
  version: 1.0.0
  contact:
    name: API Support
    email: api@nitzutz-spark.com
    url: https://nitzutz-spark.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.nitzutz-spark.com/api
    description: Production server
  - url: https://staging-api.nitzutz-spark.com/api
    description: Staging server
...
paths:
...
  /articles:
    get:
      tags:
        - Articles
      summary: List all articles
      description: Retrieve a paginated list of articles with optional filtering
      operationId: getArticles
      parameters:
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
        - name: limit
          in: query
          description: Number of items to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of items to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [createdAt, title, publishedAt]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Articles retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  articles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
    post:
      tags:
        - Articles
      summary: Create new article
      description: Add a new article to the library
      operationId: createArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleInput'
      responses:
        '201':
          description: Article created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
components:
  schemas:
    Article:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        author:
          type: string
        sourceUrl:
          type: string
          format: uri
        category:
          type: string
        publishedAt:
          type: string
          format: date-time
        imageUrl:
          type: string
          format: uri
        excerpt:
          type: string
        readTime:
          type: integer
          description: Estimated read time in minutes
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ArticleInput:
      type: object
      required:
        - title
        - content
        - category
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        content:
          type: string
          minLength: 1
        author:
          type: string
        sourceUrl:
          type: string
          format: uri
        category:
          type: string
        publishedAt:
          type: string
          format: date-time
        imageUrl:
          type: string
          format: uri
...
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication (coming soon)
security: []